Refactor alternative language title module. #206
